+++ .storybook/main.js
@@ -1,0 +1,7 @@
+module.exports = {
+  stories: ['../zoo-modules/**/*.stories.(js|mdx)'],
+  addons: [
+	'@storybook/addon-knobs',
+	'@storybook/addon-docs'
+  ],
+};