+++ rollup.config.test.js
@@ -7,9 +7,9 @@ export default [
 		input: 'src/components.js',
 		output: {
 			sourcemap: false,
-			format: 'iife',
+			format: 'esm',
 			file: 'test/bundle.js',
-			name: 'iife'
+			name: 'bundle'
 		},
 		plugins: [
 			svelte({