+++ src/App.svelte
@@ -1,7 +1,7 @@
 <div class="app">
-	<app-header></app-header>
+	<Header/>
 	<main>
-		<app-context id="what" text="What is this project?"></app-context>
+		<Context id="what" text="What is this project?"/>
 		<ul class="what-list">
 			<li>
 				Set of web-components which can be used in any modern UI framework (or without any).
@@ -12,19 +12,16 @@
 		</ul>
 		<div class="page-content">
 			<div class="overview">
-				<app-form></app-form>
+				<Form/>
 				<hr>
-				<app-buttons></app-buttons>
+				<Buttons/>
 				<hr>
-				<app-tooltip-and-feedback></app-tooltip-and-feedback>
-				<hr>
-				<app-grids></app-grids>
-				<hr>
-				<app-links></app-links>
-				<hr>
+				<Grids/>
 			</div>
 			<div id="when" class="caniuse">
-				<app-context text="When can I use it?" backbtn="{true}"></app-context>
+				<Context text="When can I use it?" backbtn="{true}"/>
+				<aside>Although safari is marked as partially supporting these features, this project does not use any features that do not work in Safari.</aside>
+				<br>
 				<div class="desktop">
 					<p class="ciu_embed" data-feature="shadowdomv1" data-periods="future_1,current,past_1,past_2" data-accessible-colours="false">
 						<a href="http://caniuse.com/#feat=shadowdomv1">Can I Use shadowdomv1?</a> Data on support for the shadowdomv1 feature across the major browsers from caniuse.com.
@@ -47,59 +44,32 @@
 				</div>
 			</div>
 			<div id="how" class="spec-docs">
-				<app-context text="How can I use it?" backbtn="{true}"></app-context>
-				<div class="left-menu">
-					{#each doclinks as link}
-						<div class="link-wrapper">
-							<a href="{link.href}">{link.text}</a>
-						</div>
-						<hr class="left-menu-separator">
-					{/each}
+				<Context text="How can I use it?" backbtn="{true}"/>
+				<div class="external-docs">
+					Documentation for each component is available at
+					<zoo-link href="https://zooplus.github.io/zoo-web-components-docs/index.html" text="Docs page" size="large" type="primary"></zoo-link>
 				</div>
-			</div>
-			<div class="content">
-				<docs-button  id="button-doc"></docs-button>
-				<hr>
-				<docs-checkbox id="checkbox-doc"></docs-checkbox>
-				<hr>
-				<docs-collapsable-list id="collapsable-list-doc"></docs-collapsable-list>
-				<hr>
-				<docs-feedback id="feedback-doc"></docs-feedback>
-				<hr>
-				<docs-footer id="footer-doc"></docs-footer>
-				<hr>
-				<docs-header id="header-doc"></docs-header>
-				<hr>
-				<docs-input id="input-doc"></docs-input>
-				<hr>
-				<docs-link id="link-doc"></docs-link>
-				<hr>
-				<docs-modal id="modal-doc"></docs-modal>
-				<hr>
-				<docs-navigation id="navigation-doc"></docs-navigation>
-				<hr>
-				<docs-radio id="radio-doc"></docs-radio>
-				<hr>
-				<docs-searchable-select id="searchable-select-doc"></docs-searchable-select>
-				<hr>
-				<docs-select id="select-doc"></docs-select>
-				<hr>
-				<docs-toast id="toast-doc"></docs-toast>
-				<hr>
-				<docs-tooltip id="tooltip-doc"></docs-tooltip>
-				<hr>
-				<docs-grid id="grid-doc"></docs-grid>
-				<hr>
-				<docs-theming id="theming-doc"></docs-theming>
-				<hr>
 			</div>
 		</div>
 	</main>
-	<zoo-footer class="footer" bind:this={footer} copyright="zooplus AG"></zoo-footer> 
+	<zoo-footer class="footer" copyright="zooplus AG">
+		{#each footerlinks as footerlink}
+			<zoo-link href="{footerlink.href}" target="{footerlink.target}" type="{footerlink.type}"
+				disabled="{footerlink.disabled}" text="{footerlink.text}">
+			</zoo-link>
+		{/each}
+	</zoo-footer> 
 </div>
 
 <style type="text/scss">
 	@import "variables";
+
+	.external-docs {
+		width: 100%;
+		display: flex;
+		align-items: center;
+		justify-content: center;
+	}
 
 	.app {
 		margin: 0 auto;
@@ -112,17 +82,13 @@
 	.page-content {
 		position: relative;
 		display: grid;
-		grid-template-columns: 340px 1fr;
+		grid-template-columns: 1fr;
 		grid-gap: 30px;
-		grid-template-areas: "overview overview"
-		"caniuse caniuse"
-		"spec-docs content";
+		grid-template-areas: "overview"
+		"caniuse"
+		"spec-docs";
 
 		@media only screen and (max-width: 850px) {
-			grid-template-areas: "overview"
-			"caniuse"
-			"spec-docs" 
-			"content";
 			grid-template-columns: minmax(320px, 90%);
 			justify-content: center;
 		}
@@ -134,6 +100,11 @@
 	}
 
 	#when {
+		aside {
+			color: var(--primary-dark, #{$primary-dark});
+			text-align: center;
+		}
+
 		.desktop {
 			@media only screen and (max-width: 850px) {
 				display: none;
@@ -175,15 +146,6 @@
 		}
 	}
 
-	.left-menu {	
-		.left-menu-separator {
-			margin: 0;
-		}
-		@media only screen and (max-width: 850px) {
-			display: none;
-		}
-	}
-
 	.overview {
 		grid-area: overview;
 		max-width: 1280px;
@@ -205,13 +167,7 @@
 
 	.spec-docs {
 		grid-area: spec-docs;
-		position: sticky;
-		top: 0;
-		height: 200px;
-	}
-
-	.content {
-		grid-area: content;
+		margin-bottom: 50px;
 	}
 
 	hr {
@@ -226,90 +182,27 @@
 </style>
 
 <script>
-	import { onMount } from 'svelte';
-	let footer;
+	import Header from './Header.svelte';
+	import Buttons from './Buttons.svelte';
+	import Form from './Form.svelte';
+	import Grids from './Grids.svelte';
+	import Context from './Context.svelte';
+	let footerlinks = [
+		{
+			href: 'https://github.com/zooplus/zoo-web-components',
+			text: 'Github',
+			type: 'negative'
+		},
+		{
+			href: 'https://www.npmjs.com/package/@zooplus/zoo-web-components',
+			text: 'NPM',
+			type: 'negative'
+		}
+	];
 	let doclinks = [
-		{
-			href: '#button-doc',
-			text: 'Button'
-		},
-		{
-			href: '#checkbox-doc',
-			text: 'Checkbox'
-		},
-		{
-			href: '#collapsable-list-doc',
-			text: 'Collapsable List'
-		},
-		{
-			href: '#feedback-doc',
-			text: 'Feedback'
-		},
-		{
-			href: '#footer-doc',
-			text: 'Footer'
-		},
-		{
-			href: '#header-doc',
-			text: 'Header'
-		},
-		{
-			href: '#input-doc',
-			text: 'Input'
-		},
-		{
-			href: '#link-doc',
-			text: 'Link'
-		},
-		{
-			href: '#modal-doc',
-			text: 'Modal'
-		},
-		{
-			href: '#navigation-doc',
-			text: 'Navigation'
-		},
-		{
-			href: '#radio-doc',
-			text: 'Radio'
-		},
-		{
-			href: '#searchable-select-doc',
-			text: 'Searchable select'
-		},
-		{
-			href: '#select-doc',
-			text: 'Select'
-		},
-		{
-			href: '#toast-doc',
-			text: 'Toast'
-		},
-		{
-			href: '#tooltip-doc',
-			text: 'Tooltip'
-		},
-		{
-			href: '#grid-doc',
-			text: 'Grid'
-		},
 		{
 			href: '#theming-doc',
 			text: 'Theming'
 		}
 	];
-	onMount(() => {
-		footer.footerlinks = [
-			{
-				href: 'https://github.com/zooplus/zoo-web-components',
-				text: 'Github',
-				type: 'negative'
-			},
-			{
-				href: 'https://www.npmjs.com/package/@zooplus/zoo-web-components',
-				text: 'NPM',
-				type: 'negative'
-			}
-		];
-	});
 </script>