+++ zoo-modules/header-module/Header.svelte
@@ -1,7 +1,9 @@
 <svelte:options tag="zoo-header"></svelte:options>
-<header bind:this={_headerRoot}>
-	<img importance="high" src="{imgsrc}" alt="{imgalt}" bind:this={_img}/>
-	<h2>{headertext}</h2>
+<header>
+	<slot name="img"></slot>
+	<slot name="headertext">
+		<h2>{headertext}</h2>
+	</slot>
 	<slot></slot>
 </header>
 
@@ -20,7 +22,7 @@
 		height: 70px;
 	}
 
-	img {
+	::slotted(img) {
 		height: 46px;
 		display: inline-block;
 		padding: 5px 25px 5px 0;
@@ -30,7 +32,7 @@
 		}
 	}
 
-	h2 {
+	::slotted(*[slot="headertext"]), h2 {
 		display: inline-block;
 		color: var(--primary-mid, #{$primary-mid});
 		@media only screen and (max-width: 544px) {
@@ -40,16 +42,5 @@
 </style>
 
 <script>
-	import { onMount } from 'svelte';
 	export let headertext = '';
-	export let imgsrc = '';
-	export let imgalt = '';
-	let _headerRoot;
-	let _img;
-
-	onMount(() => {
-		const host = _headerRoot.getRootNode().host;
-		_img.addEventListener("click", () => host.dispatchEvent(new Event("logoClicked")));
-	});
-
 </script>