+++ zoo-modules/shared/groups.js
@@ -1,0 +1,2 @@
+export const attributesGroupId = 'Attributes';
+export const cssVariablesGroupId = 'CSS variables';